+++ src/components/navigation-page/index.js
@@ -1,0 +1,3 @@
+import main from './navigation-page';
+
+export default main;