+++ src/components/page-layout/layout.config.js
@@ -5,12 +5,10 @@ export default {
   },
   accordion: {
     sidebar: 'accordion', // NavigationAccordion sidebar
-    sidebarTheme: 'bg-gray-faint' // sidebar background
   },
   example: {
     sidebar: 'sectioned', // SectionedNavigation sidebar
     hideSubItems: false, // show headings and subitems in sitebar
-    sidebarTheme: 'bg-gray-faint', // sidebar background
     includeFilterBar: false // hide filter bar
   },
   exampleIndex: {
@@ -18,7 +16,6 @@ export default {
     hideSubItems: true, // only show sidebar headings
     showCards: true, // show example cards
     hideFeedback: true, // hide feedback module
-    sidebarTheme: 'bg-gray-faint', // sidebar background
     includeFilterBar: false // hide filter bar
   },
   full: {